Changed int to size_t

+++ io/reader.hpp
@@ -11,12 +11,12 @@ namespace io {
 
   template<class V>
   struct TABReader {
-    void Read(istream &is, int *num_objects, int *num_dimensions, int *num_clusters, vector<V> *objects) {
+    void Read(istream &is, size_t *num_objects, size_t *num_dimensions, size_t *num_clusters, vector<V> *objects) {
       is >> *num_objects >> *num_dimensions >> *num_clusters;
 
       objects->resize(*num_objects, V(*num_objects));
-      for (int i = 0; i < *num_objects; ++i)
-	for (int j = 0; j < *num_dimensions; ++j)
+      for (size_t i = 0; i < *num_objects; ++i)
+	for (size_t j = 0; j < *num_dimensions; ++j)
 	  is >> (*objects)[i][j];
     }
   }; // struct TABReader
